requirements.txt
Django==1.6.1
South==0.8.1
amqp==1.0.13
ansi2html==0.9.3
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.32
celery==3.0.23
django-celery==3.0.23
django-crispy-forms==1.4.0
django-redis==3.3
django-tastypie==0.9.16
-e git+https://github.com/shipyard/docker-py@master#egg=docker_py
kombu==2.5.14
mimeparse==0.1.3
python-dateutil==2.1
pytz==2013b
redis==2.7.6
requests==1.2.3
six==1.3.0
times==0.6.1
wsgiref==0.1.2
django-json-field==0.5.5
ipython
cloud
